fix: fix logic bug in several Member Controller (IDOR vulnerability issues) ; adjust responsiveness for sidebard (cart & order preview in admin/owner dashboard)

+++ app/Http/Controllers/Member/TestimonyController.php
@@ -109,7 +109,11 @@ class TestimonyController extends Controller
 
         $date = now();
 
-        $testimony = Testimony::where('product_id', $id)->first();
+        $testimony = Testimony::where('product_id', $id)->where('user_id', Auth::id())->first();
+
+        if (!$testimony) {
+            abort(403, 'Ulasan tidak ditemukan');
+        }
 
         if ($request->file('image')) {
             if ($request->oldImage) {
@@ -139,7 +143,10 @@ class TestimonyController extends Controller
      */
     public function destroy($id)
     {
-        $testimony = Testimony::where('id', $id)->first();
+        $testimony = Testimony::where('id', $id)->where('user_id', Auth::id())->first();
+        if (!$testimony) {
+            abort(403, 'Ulasan tidak ditemukan');
+        }
         if ($testimony->image != null) {
             Storage::disk('public')->delete($testimony->image);
         }
